# Tweak hot paths