@lahma 2018-04-25
I've changed the PR to contain the changes that still seem relevant after rebase with dev. Some modest improvements.